neew

+++ Kool/Controllers/KoolitusController.cs
@@ -7,6 +7,7 @@ using System.Net;
 using System.Web;
 using System.Web.Mvc;
 using Kool.Models;
+using Microsoft.AspNet.Identity;
 
 namespace Kool.Controllers
 {
@@ -36,6 +37,7 @@ namespace Kool.Controllers
         }
 
         // GET: Koolitus/Create
+        [Authorize(Roles = "Admin")]
         public ActionResult Create()
         {
             return View();
@@ -46,6 +48,7 @@ namespace Kool.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = "Admin")]
         public ActionResult Create([Bind(Include = "Id,KeelekursusId,OpetajaId,AlgusKuupaev,LoppKuupaev,Hind,MaxOsalejaid")] Koolitus koolitus)
         {
             if (ModelState.IsValid)
@@ -59,6 +62,7 @@ namespace Kool.Controllers
         }
 
         // GET: Koolitus/Edit/5
+        [Authorize(Roles = "Admin")]
         public ActionResult Edit(int? id)
         {
             if (id == null)
@@ -78,6 +82,7 @@ namespace Kool.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = "Admin")]
         public ActionResult Edit([Bind(Include = "Id,KeelekursusId,OpetajaId,AlgusKuupaev,LoppKuupaev,Hind,MaxOsalejaid")] Koolitus koolitus)
         {
             if (ModelState.IsValid)
@@ -90,6 +95,7 @@ namespace Kool.Controllers
         }
 
         // GET: Koolitus/Delete/5
+        [Authorize(Roles = "Admin")]
         public ActionResult Delete(int? id)
         {
             if (id == null)
@@ -103,13 +109,30 @@ namespace Kool.Controllers
             }
             return View(koolitus);
         }
+        [Authorize(Roles = "Opetaja")]
+        public ActionResult MinuKoolitused()
+        {
+            string userId = User.Identity.GetUserId();
+
+            var minu = db.Koolitused
+                .Include(k => k.Keelekursus)
+                .Include(k => k.Opetaja)
+                .Where(k => k.Opetaja.ApplicationUserId == userId)
+                .OrderByDescending(k => k.AlgusKuupaev)
+                .ToList();
+
+            return View(minu);
+        }
 
         // POST: Koolitus/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = "Admin")]
         public ActionResult DeleteConfirmed(int id)
         {
-            Koolitus koolitus = db.Koolitused.Find(id);
+            var koolitus = db.Koolitused.Find(id);
+            if (koolitus == null) return HttpNotFound();
+
             db.Koolitused.Remove(koolitus);
             db.SaveChanges();
             return RedirectToAction("Index");
